+++ src/pages/AdicionarPratoPrincipal/index.js
@@ -26,6 +26,7 @@ export default function AdicionarPratoPrincipal({ route, navigation }) {
     const [isSugestao, setIsSugestao] = React.useState(false);
     const [isFirstTime, setIsFirstTime] = React.useState(primeiroAcesso);
     const [isEnabled, setIsEnabled] = useState(false);
+    const [modalSair, setModalSair] = useState(false)
 
     const config = {
         headers: { 'Authorization': USUARIOLOGADO.id }
@@ -75,16 +76,30 @@ export default function AdicionarPratoPrincipal({ route, navigation }) {
         navigation.push('EscolherNovosItens', { churrascode })
     }
 
-    function backHome() {
-        LISTADECONVIDADOS = null;
-        CONVITE = null;
-        setLoading(true)
-        api.delete(`/churras/${churrascode}`, config)
+    
+    function backHome(sair, ficar) {
+        if(sair === true) {
+            LISTADECONVIDADOS = null;
+            CONVITE = null;
+            setLoading(true)
+            api.delete(`/churras/${churrascode}`, config)
             .then(function () {
                 setLoading(false)
                 navigation.replace('Tabs');
             })
-    }
+          setModalSair(false)
+        }
+        if(ficar === true){
+          setModalSair(false)
+        }
+       }
+    
+       function backOnePage(){
+        setLoading(true)
+        setItemList([])
+        setLoading(false)
+        navigation.goBack()
+   }
 
     function updateValue(qtdSugestao) {
         if (isSugestao) {
@@ -149,11 +164,16 @@ export default function AdicionarPratoPrincipal({ route, navigation }) {
         <View style={style.container}>
             <SafeAreaView style={style.body}>
                 <View style={style.headerGroup}>
+                <TouchableOpacity style={style.exitBtn} onPress={() => backOnePage()}>
+                        <Icon style={style.iconHeaderBtn} name="md-arrow-back" size={22} />
+                    </TouchableOpacity>
                     <View style={style.headerTextGroup}>
+                    
                         <Text style={style.textHeader}>Carnes:</Text>
                     </View>
-                    <TouchableOpacity style={style.exitBtn} onPress={() => backHome()}>
-                        <Icon style={style.iconHeaderBtn} name="md-exit" size={22} />
+                    
+                    <TouchableOpacity style={style.exitBtn} onPress={() => setModalSair(true)}>
+                        <Icon style={style.iconHeaderBtn} name="md-close" size={22} />
                     </TouchableOpacity>
                 </View>
                 {isSugestao
@@ -250,6 +270,27 @@ export default function AdicionarPratoPrincipal({ route, navigation }) {
                         <Text style={style.textBtn}>Continuar</Text>
                     </TouchableOpacity>
                 </View>
+        <Modal
+          animationType="slide"
+          transparent={true}
+          visible={modalSair}
+        >
+          <View style={style.centeredView}>
+            <View style={style.modalView}>
+              <Text style={style.modalTitle}>Quer sair?</Text>
+              <Text style={style.modalText}>Você deseja mesmo desfazer este churras?</Text>
+              <Text style={style.confirmarSairSubTitle}>(Ele sera completamente perdido, mas nunca esquecido)</Text>
+              <View style={style.footerModal}>
+                 <TouchableOpacity style={style.sairBtn} onPress={() => backHome(false, true)}>
+                  <Text style={style.textBtn}>Não</Text>
+                </TouchableOpacity>
+                <TouchableOpacity style={style.sairBtn} onPress={() => backHome(true, false)}>
+                  <Text style={style.textBtn}>Claro</Text>
+                </TouchableOpacity>
+              </View>
+            </View>
+          </View>
+        </Modal>
                 {criarModal}
             </SafeAreaView>
         </View>
